Improve fn_80156DA0 via int randBit (drop spurious clrlwi)
randBit was u16, forcing a clrlwi mask before the (randBit-1)*0x12c
arithmetic; the target keeps it as int. T=175->174 instrs, 8->7 regions.

+++ src/main/dll/duster.c
@@ -1046,7 +1046,7 @@ void fn_80156DA0(int obj, int state)
     bool resetting;
     int groundHit;
     u8 noHit;
-    u16 randBit;
+    int randBit;
     float toPos[3];
     float fromPos[3];
     float cosYaw;
